requirements.txt
# Core API Server
fastapi
uvicorn
pydantic

# LLM & AI (OpenRouter uses OpenAI SDK)
openai>=1.0.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Langchain for Memory Management and Query Refinement
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Vector Database
faiss-cpu>=1.7.4  # Sử dụng faiss-gpu nếu có card đồ họa NVIDIA

# Data Processing
numpy>=1.21.0
tqdm>=4.62.0
python-dotenv

# Vietnamese NLP (Optional)
underthesea>=1.3.5
pyvi>=0.1.1
